+++ memberaudit/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "memberaudit.apps.MemberAuditConfig"
 
-__version__ = "1.5.1"
+__version__ = "1.6.0a1"
 __title__ = "Member Audit"